Backend/requirements.txt
ibm-watsonx-ai>=1.1.9
python-dotenv>=1.0.1
flask
flask-cors
flask-sock
pyserial>=3.5